Code Refactoring: Separation of DEBUG Code Paths into DebuggingSceneDelegate Subclas.
To enhance code separation between production and debugging/testing environments, the DEBUG code paths previously located in the main SceneDelegate have been migrated to a new subclass called DebuggingSceneDelegate. This move ensures that debug- and test-specific code is isolated from the production code.
