+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -10,11 +10,19 @@ import EssentialFeed
 import EssentialFeediOS
 import UIKit
 
-final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
 
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+    let localStoreURL = NSPersistentContainer
+        .defaultDirectoryURL()
+        .appendingPathComponent("feed-store.sqlite")
+
+    func scene(
+        _ scene: UIScene,
+        willConnectTo session: UISceneSession,
+        options connectionOptions: UIScene.ConnectionOptions
+    ) {
         guard let _ = (scene as? UIWindowScene) else { return }
 
         let remoteURL = URL(string: "https://static1.squarespace.com/static/5891c5b8d1758ec68ef5dbc2/t/5db4155a4fbade21d17ecd28/1572083034355/essential_app_feed.json")!
@@ -26,12 +34,6 @@ final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let localStoreURL = NSPersistentContainer
             .defaultDirectoryURL()
             .appendingPathComponent("feed-store.sqlite")
-
-        #if DEBUG
-            if CommandLine.arguments.contains("-reset") {
-                try? FileManager.default.removeItem(at: localStoreURL)
-            }
-        #endif
 
         let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
         let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -50,26 +52,7 @@ final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
                     cache: localImageLoader)))
     }
 
-    private func makeRemoteClient() -> HTTPClient {
-        #if DEBUG
-            if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-                return AlwaysFailingHTTPClient()
-            }
-        #endif
-
-        return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
+    func makeRemoteClient() -> HTTPClient {
+        URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
     }
 }
-
-#if DEBUG
-    private class AlwaysFailingHTTPClient: HTTPClient {
-        final class Task: HTTPClientTask {
-            func cancel() {}
-        }
-
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            completion(.failure(NSError(domain: "offline", code: 0)))
-            return Task()
-        }
-    }
-#endif
